Update user db model and refactor schemas
