--- backend/app/schemas/__init__.py
@@ -1,2 +1,0 @@
-from . import routers  # noqa: F401
-from . import db  # noqa: F401
