+++ backend/app/schemas/chat.py
@@ -1,0 +1,16 @@
+from typing import List, Literal
+from pydantic import BaseModel, constr
+
+
+class ChatMessage(BaseModel):
+    role: Literal["user", "assistant", "system"]
+    content: constr(strip_whitespace=True, min_length=1)
+
+
+class ChatRequest(BaseModel):
+    messages: List[ChatMessage]
+    mode: str = "default"
+
+
+class ChatResponse(BaseModel):
+    reply: str
