+++ backend/app/services/classification.py
@@ -5,8 +5,12 @@ from typing import Any, Dict, List, Literal
 
 # Pydantic/Data models
 from binary_classifiers.predict_class import PredictClass
-from ..schemas.routers import ModelConfig
-from ..schemas.db import SequenceInput, SequenceResult, ClassificationResponse
+from ..schemas.classification import (
+    ModelConfig,
+    SequenceInput,
+    SequenceResult,
+    ClassificationResponse,
+)
 
 # Helpers
 from ..utils.dna_validation import validate_dna_sequence
